* debug OpenMPI installation

* add Dockerfile example, update the script

* finetune changes

+++ horovod_install.sh
@@ -15,6 +15,7 @@
 #
 # It is required that:
 # cmake
+# git
 # wget
 # libnccl2 and libnccl-dev
 # are installed.
@@ -47,8 +48,6 @@ elif [[ $1 == "-h" ]] || [[ $1 == "--help" ]]; then
     exit 1
 else 
     # read options as parameters (1)
-    #cmd=${arr[0]} # command
-    #params=${arr[@]:1}
     params="$*"
 fi
 
@@ -81,19 +80,23 @@ if [ ${#OpenMPI} -gt 2 ]; then
     if $OpenMPI_Install; then
         # deduce the main OpenMPI version
         OpenMPI_MainVer=$(echo ${OpenMPI} | cut -d\. -f1,2)
-        mkdir /tmp/openmpi && \
-        cd /tmp/openmpi && \
-        wget "https://www.open-mpi.org/software/ompi/v${OpenMPI_MainVer}/downloads/openmpi-${OpenMPI}.tar.gz" && \
-        tar zxf openmpi-${OpenMPI}.tar.gz && \
-        cd openmpi-${OpenMPI} && \
-        ./configure --enable-orterun-prefix-by-default && \
-        make -j $(nproc) all && \
-        make install && \
-        ldconfig && \
-        rm -rf /tmp/openmpi
+        mkdir /tmp/openmpi && cd /tmp/openmpi && \
+        wget "https://www.open-mpi.org/software/ompi/v${OpenMPI_MainVer}/downloads/openmpi-${OpenMPI}.tar.gz"
+        tar zxf openmpi-${OpenMPI}.tar.gz
+        cd openmpi-${OpenMPI}
+        ./configure --enable-orterun-prefix-by-default
+        [[ $? -eq 0 ]] && make -j $(nproc) all
+        [[ $? -eq 0 ]] && make install
+        [[ $? -eq 0 ]] && ldconfig
+        [[ $? -eq 0 ]] && rm -rf /tmp/openmpi
+        echo "[DEBUG] Finish OpenMPI installation"
     fi
 fi
 
+# go back to the script folder
+cd ${SCRIPT_PATH}
+
+# install HOROVOD if asked for
 if [ ${#HOROVOD} -gt 2 ]; then
     if [ ${HOROVOD}=="latest" ]; then
         HOROVOD_PYPI="horovod"
@@ -103,12 +106,20 @@ if [ ${#HOROVOD} -gt 2 ]; then
     # Install Horovod, temporarily using CUDA stubs (Ubuntu path!)
     # pip3 also checks, if horovod is already installed
     ldconfig /usr/local/cuda/targets/x86_64-linux/lib/stubs && \
-    HOROVOD_GPU_ALLREDUCE=NCCL HOROVOD_WITH_TENSORFLOW=1 pip3 install --no-cache-dir ${HOROVOD_PYPI} && \
+    HOROVOD_GPU_ALLREDUCE=NCCL HOROVOD_WITH_TENSORFLOW=1 \
+    pip3 install --no-cache-dir ${HOROVOD_PYPI} && \
     ldconfig
 fi
 
-# command to execute at the end
+# go back to the script folder
 cd ${SCRIPT_PATH}
+
+# command to execute at the end
+if [ $? -eq 0 ]; then
 /bin/bash<<EOF
 ${params}
 EOF
+else
+    echo "[ERROR] Something went wrong. Please, check error messages above"
+fi
+
